@@ -1,44 +1,44 @@
 {
-  "name": "cpsit/auditor",
-  "description": "Audit PHP application bundles.",
-  "license": "GPL-2.0-or-later",
-  "type": "composer-plugin",
-  "require": {
-    "php": "^7.4 || ^8.0 ||^8.1",
-    "composer-plugin-api": "^2.0.0"
-  },
-  "authors": [
-    {
-      "name": "Dirk Wenzel",
-      "email": "dirk.wenzel@cps-it.de"
-    }
-  ],
-  "config": {
-    "vendor-dir": ".build/vendor",
-    "bin-dir": ".build/bin"
-  },
-  "scripts": {
-    "post-autoload-dump": [
-      "mkdir -p .build/log/coverage/"
-    ],
-    "test": "phpunit -c tests/Build/UnitTests.xml"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "^7.5.0 || ^8.0.0 || ^9.0.0",
-    "composer/composer": "^2.1.9",
-    "symfony/filesystem": "~4.4.0 || ~5.2.0"
-  },
-  "autoload": {
-    "psr-4": {
-      "CPSIT\\Auditor\\": "src"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "CPSIT\\Auditor\\Tests\\": "tests/"
-    }
-  },
-  "extra": {
-    "class": "CPSIT\\Auditor\\Installer"
-  }
+	"name": "cpsit/auditor",
+	"description": "Audit PHP application bundles.",
+	"license": "GPL-2.0-or-later",
+	"type": "composer-plugin",
+	"require": {
+		"php": "^8.2 || ^8.3",
+		"composer-plugin-api": "^2.1"
+	},
+	"authors": [
+		{
+			"name": "Dirk Wenzel",
+			"email": "dirk.wenzel@cps-it.de"
+		}
+	],
+	"config": {
+		"vendor-dir": ".build/vendor",
+		"bin-dir": ".build/bin"
+	},
+	"scripts": {
+		"post-autoload-dump": [
+			"mkdir -p .build/log/coverage/"
+		],
+		"test": "phpunit -c tests/Build/UnitTests.xml"
+	},
+	"require-dev": {
+		"phpunit/phpunit": "^10.5 || ^11.0",
+		"composer/composer": "^2.1",
+		"symfony/filesystem": ">=6.4"
+	},
+	"autoload": {
+		"psr-4": {
+			"CPSIT\\Auditor\\": "src"
+		}
+	},
+	"autoload-dev": {
+		"psr-4": {
+			"CPSIT\\Auditor\\Tests\\": "tests/"
+		}
+	},
+	"extra": {
+		"class": "CPSIT\\Auditor\\Installer"
+	}
 }
